create abstract class MealSuggesterView and refactor HighCalorieMealsView, KetoSuggestionHelperView, SweetWithoutEggView
